+++ lets/__init__.py
@@ -1,6 +1,6 @@
 from lets import module
 
-def do(path:str, data:bytes, options:dict=None) -> bytes:
+def do(path:str, data:bytes=None, options:dict=None, generate:bool=False) -> bytes:
     """
     Build and execute a module with the specified data
     and options.
@@ -8,11 +8,12 @@ def do(path:str, data:bytes, options:dict=None) -> bytes:
     :param path: Path to module
     :param data: Data to be used by module, in bytes
     :param options: Dict of options to be used by module
+    :param generate: Choose to return a generator or bytes
     :return: Results of module execution, in bytes
     """
     mod = module.Module.build(path)
     if mod:
         gen = mod.do(data, options)
         if gen:
-            return b"".join(gen)
-    return b""
+            return gen if generate else b"".join(gen)
+    return iter(()) if generate else b""