+++ lets/modules/sample/mymodule.py
@@ -29,7 +29,7 @@ class MyModule(Module):
 
         return parser
 
-    def do(self, data:bytes, options:dict=None) -> bytes:
+    def do(self, data:bytes=None, options:dict=None) -> bytes:
         """
         Main functionality.