+++ test.py
@@ -1,6 +1,50 @@
 #!/usr/bin/env python3
-import os, sys, argparse, unittest
-from lets.module import Module
+import os, sys, argparse, unittest, subprocess
+import lets
+
+class BashInterfaceTest(unittest.TestCase):
+    """
+    Test the bash interface.  Relies on the following modules:
+
+    ["encode/base64"]
+    """
+    def test(self):
+        """
+        Perform unit tests to verify the bash interface functionality.
+        """
+        _lets = os.path.sep.join([os.path.abspath(os.path.dirname(__file__)), "lets.py"])
+
+        p = subprocess.Popen(["python3", _lets, "encode/base64"],
+            stdin=subprocess.PIPE, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+        out,err = p.communicate(input=b"abcd")
+        self.assertEqual(
+            out,
+            b"YWJjZA==",
+            "Base64 encoding produced inaccurate results")
+
+
+
+class PythonInterfaceTest(unittest.TestCase):
+    """
+    Test the python interface.  Relies on the following modules:
+
+    ["encode/base64"]
+    """
+    def test(self):
+        """
+        Perform unit tests to verify the python interface functionality.
+        """
+
+        self.assertEqual(
+            lets.do("encode/base64", b"abcd", {"verbose" : False}),
+            b"YWJjZA==",
+            "Base64 encoding produced inaccurate bytes")
+
+        self.assertEqual(
+            b"".join(lets.do("encode/base64", b"abcd", {"verbose" : False}, generate=True)),
+            b"YWJjZA==",
+            "Base64 encoding produced inaccurate generator")
+
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser(
@@ -15,16 +59,20 @@ if __name__ == "__main__":
 
     if args.module:
         # Build module and add test
-        mod = Module.build(args.module)
+        mod = lets.module.Module.build(args.module)
         if mod:
             suite.addTest(mod)
         else:
-            raise(Module.Exception("Error loading module: %s" % args.module))
+            raise(lets.module.Module.Exception("Error loading module: %s" % args.module))
 
     else:
+        # Add interface tests
+        suite.addTest(BashInterfaceTest(lets.module.Module.test_method))
+        suite.addTest(PythonInterfaceTest(lets.module.Module.test_method))
+
         # Walk modules and add tests
-        for mod in Module.build_all():
-            if mod and hasattr(mod, Module.test_method):
+        for mod in lets.module.Module.build_all():
+            if mod and hasattr(mod, lets.module.Module.test_method):
                 suite.addTest(mod)
 
     # Execute test suite